Move GPT-2 in the model catalog
