+++ opennmt/layers/common.py
@@ -1,6 +1,7 @@
 """Defines common layers."""
 
 import tensorflow as tf
+import numpy as np
 
 from opennmt.utils.misc import shape_list
 
@@ -10,6 +11,12 @@ def dropout(x, rate, training=None):
   if not training or rate == 0:
     return x
   return tf.nn.dropout(x, rate)
+
+def gelu(x):
+  """Gaussian Error Linear Unit activation function described in
+  https://arxiv.org/abs/1606.08415.
+  """
+  return 0.5 * x * (1 + tf.tanh(np.sqrt(2 / np.pi) * (x + 0.044715 * tf.pow(x, 3))))
 
 
 class Dense(tf.keras.layers.Dense):
